requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=1.26.0,<3.0.0